move arbitrary js inside the createExecutionScope
