+++ editor/src/components/canvas/ui-jsx-canvas-renderer/ui-jsx-canvas-execution-scope.ts
@@ -1,16 +1,26 @@
 import * as React from 'react'
 import type { MapLike } from 'typescript'
-import { isUtopiaJSXComponent, UtopiaJSXComponent } from '../../../core/shared/element-template'
+import {
+  ArbitraryJSBlock,
+  isUtopiaJSXComponent,
+  UtopiaJSXComponent,
+} from '../../../core/shared/element-template'
 import { fastForEach } from '../../../core/shared/utils'
 import { ProjectContentTreeRoot } from '../../assets'
 import { importResultFromImports } from '../../editor/npm-dependency/npm-dependency'
-import { TransientFileState } from '../../editor/store/editor-state'
+import { TransientFileState, UIFileBase64Blobs } from '../../editor/store/editor-state'
 import {
   ComponentRendererComponent,
   createComponentRendererComponent,
 } from './ui-jsx-canvas-component-renderer'
 import { MutableUtopiaContextProps } from './ui-jsx-canvas-contexts'
+import { createLookupRender, utopiaCanvasJSXLookup } from './ui-jsx-canvas-element-renderer-utils'
+import { runBlockUpdatingScope } from './ui-jsx-canvas-scope-utils'
 import { getTopLevelElements } from './ui-jsx-canvas-top-level-elements'
+import * as TP from '../../../core/shared/template-path'
+import { UiJsxCanvasContextData } from '../ui-jsx-canvas'
+import { TemplatePath } from '../../../core/shared/project-file-types'
+import { JSX_CANVAS_LOOKUP_FUNCTION_NAME } from '../../../core/workers/parser-printer/parser-printer-utils'
 
 export function createExecutionScope(
   filePath: string,
@@ -22,6 +32,12 @@ export function createExecutionScope(
   projectContents: ProjectContentTreeRoot,
   openStoryboardFileNameKILLME: string | null,
   transientFileState: TransientFileState | null,
+  combinedTopLevelArbitraryBlock: ArbitraryJSBlock | null,
+  fileBlobs: UIFileBase64Blobs, // this is only for the storyboard file now,
+  hiddenInstances: TemplatePath[],
+  metadataContext: UiJsxCanvasContextData,
+  jsxFactoryFunction: string | null,
+  shouldIncludeCanvasRootInTheSpy: boolean,
 ): {
   scope: MapLike<any>
   topLevelJsxComponents: Map<string, UtopiaJSXComponent>
@@ -63,6 +79,32 @@ export function createExecutionScope(
       }
     }
   })
+
+  // First make sure everything is in scope
+  if (combinedTopLevelArbitraryBlock != null && openStoryboardFileNameKILLME != null) {
+    const lookupRenderer = createLookupRender(
+      TP.emptyInstancePath,
+      executionScope,
+      {},
+      requireResult,
+      hiddenInstances,
+      fileBlobs,
+      [],
+      undefined,
+      metadataContext,
+      jsxFactoryFunction,
+      shouldIncludeCanvasRootInTheSpy,
+      openStoryboardFileNameKILLME,
+    )
+
+    executionScope[JSX_CANVAS_LOOKUP_FUNCTION_NAME] = utopiaCanvasJSXLookup(
+      combinedTopLevelArbitraryBlock.elementsWithin,
+      executionScope,
+      lookupRenderer,
+    )
+
+    runBlockUpdatingScope(requireResult, combinedTopLevelArbitraryBlock, executionScope)
+  }
 
   return {
     scope: {
